+++ config/config.php
@@ -15,7 +15,7 @@
 /**
  * Extension version
  */
-@define('ISOTOPE_PAYU_VERSION', '1.0');
+@define('ISOTOPE_PAYU_VERSION', '1.1');
 @define('ISOTOPE_PAYU_BUILD', '1');