+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 isotope_payu Changelog
 ======================
 
-Version 1.0.1 stable (2012-01-26)
+Version 1.1.1 stable (2013-01-28)
+---------------------------------
+
+### Improved
+- Added the backend order interface
+- Removed old code
+- Changed the payment description
+
+
+Version 1.0.1 stable (2013-01-26)
 ---------------------------------
 
 ### Fixed
@@ -11,7 +20,7 @@ Version 1.0.1 stable (2012-01-26)
 - Extended the list of available status options
 
 
-Version 1.0.0 stable (2012-01-20)
+Version 1.0.0 stable (2013-01-20)
 ---------------------------------
 
 Initial release.